+++ substrate/frame/support/procedural/src/pallet/expand/call.rs
@@ -17,12 +17,14 @@
 
 use crate::{
 	pallet::{
+		expand::warnings::{weight_constant_warning, weight_witness_warning},
 		parse::call::{CallVariantDef, CallWeightDef},
 		Def,
 	},
 	COUNTER,
 };
 use proc_macro2::TokenStream as TokenStream2;
+use proc_macro_warning::Warning;
 use quote::{quote, ToTokens};
 use syn::spanned::Spanned;
 
@@ -68,7 +70,7 @@ pub fn expand_call(def: &mut Def) -> proc_macro2::TokenStream {
 			continue
 		}
 
-		let warning = proc_macro_warning::Warning::new_deprecated("ImplicitCallIndex")
+		let warning = Warning::new_deprecated("ImplicitCallIndex")
 			.index(call_index_warnings.len())
 			.old("use implicit call indices")
 			.new("ensure that all calls have a `pallet::call_index` attribute or put the pallet into `dev` mode")
@@ -86,20 +88,11 @@ pub fn expand_call(def: &mut Def) -> proc_macro2::TokenStream {
 	for method in &methods {
 		match &method.weight {
 			CallWeightDef::DevModeDefault => fn_weight.push(syn::parse_quote!(0)),
-			CallWeightDef::Immediate(e @ syn::Expr::Lit(lit)) if !def.dev_mode => {
-				let warning = proc_macro_warning::Warning::new_deprecated("ConstantWeight")
-					.index(weight_warnings.len())
-					.old("use hard-coded constant as call weight")
-					.new("benchmark all calls or put the pallet into `dev` mode")
-					.help_link("https://github.com/paritytech/substrate/pull/13798")
-					.span(lit.span())
-					.build_or_panic();
-				weight_warnings.push(warning);
-				fn_weight.push(e.into_token_stream());
-			},
 			CallWeightDef::Immediate(e) => {
-				fn_weight.push(e.clone().into_token_stream());
-				check_weight_witness(method, &mut weight_warnings);
+				weight_constant_warning(e, def.dev_mode, &mut weight_warnings);
+				weight_witness_warning(method, &mut weight_warnings);
+
+				fn_weight.push(e.into_token_stream());
 			},
 			CallWeightDef::Inherited => {
 				let pallet_weight = def
@@ -427,47 +420,4 @@ pub fn expand_call(def: &mut Def) -> proc_macro2::TokenStream {
 			}
 		}
 	)
-}
-
-/// Warn if any of the call arguments starts with a underscore and is used in a weight formula.
-fn check_weight_witness(method: &CallVariantDef, warnings: &mut Vec<proc_macro_warning::Warning>) {
-	let CallWeightDef::Immediate(imm) = &method.weight else {
-		return;
-	};
-	let builder = proc_macro_warning::Warning::new_deprecated("UncheckedWeightWitness")
-		.old("not checked weight witness data")
-		.new("ensure that all witness data for weight calculation is checked before usage")
-		.help_links(&["FAIL-CI TODO"]);
-
-	for (_, arg_ident, _) in method.args.iter().skip(1) {
-		// Unused arguments cannot be used in weight formulas.
-		if !arg_ident.to_string().starts_with('_') || !contains(&imm, &arg_ident) {
-			continue
-		}
-
-		let warning = builder.clone().index(warnings.len()).span(arg_ident.span()).build_or_panic();
-		warnings.push(warning);
-	}
-}
-
-fn contains(expr: &syn::Expr, ident: &syn::Ident) -> bool {
-	use syn::visit_mut::{self, VisitMut};
-
-	struct ContainsIdent {
-		ident: syn::Ident,
-		found: bool,
-	}
-
-	impl VisitMut for ContainsIdent {
-		fn visit_ident_mut(&mut self, i: &mut syn::Ident) {
-			if *i == self.ident {
-				self.found = true;
-			}
-		}
-	}
-
-	let mut expr = expr.clone();
-	let mut visitor = ContainsIdent { ident: ident.clone(), found: false };
-	visit_mut::visit_expr_mut(&mut visitor, &mut expr);
-	visitor.found
 }
